+++ src/main/java/co/fineants/api/domain/member/domain/entity/MemberProfile.java
@@ -3,7 +3,7 @@ package co.fineants.api.domain.member.domain.entity;
 import java.util.Map;
 import java.util.Optional;
 
-import co.fineants.api.domain.member.domain.rule.ValidationRule;
+import co.fineants.api.domain.validator.MemberValidationRule;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
@@ -68,11 +68,11 @@ public class MemberProfile {
 		return Optional.ofNullable(profileUrl);
 	}
 
-	public void validateEmail(ValidationRule rule) {
+	public void validateEmail(MemberValidationRule rule) {
 		rule.validate(email);
 	}
 
-	public void validateNickname(ValidationRule rule) {
+	public void validateNickname(MemberValidationRule rule) {
 		rule.validate(nickname);
 	}
 }
