--- src/main/java/co/fineants/api/domain/member/domain/rule/NicknameValidator.java
@@ -1,19 +1,0 @@
-package co.fineants.api.domain.member.domain.rule;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class NicknameValidator implements Validator<String> {
-
-	private final List<ValidationRule> rules;
-
-	public NicknameValidator(ValidationRule... rules) {
-		this.rules = Arrays.asList(rules);
-	}
-
-	public void validate(String nickname) {
-		for (ValidationRule rule : rules) {
-			rule.validate(nickname);
-		}
-	}
-}
