+++ src/main/java/co/fineants/api/domain/validator/MemberValidationRule.java
@@ -1,8 +1,8 @@
-package co.fineants.api.domain.member.domain.rule;
+package co.fineants.api.domain.validator;
 
 import co.fineants.api.domain.member.domain.entity.Member;
 
-public interface ValidationRule {
+public interface MemberValidationRule {
 	void validate(String text);
 
 	void validate(Member member);
