+++ src/main/java/co/fineants/api/domain/validator/Validator.java
@@ -1,4 +1,4 @@
-package co.fineants.api.domain.member.domain.rule;
+package co.fineants.api.domain.validator;
 
 public interface Validator<T> {
 	void validate(T target);
